+++ src/notsotiny/sim/memory/MemoryManager.java
@@ -10,11 +10,11 @@ import java.util.TreeSet;
  */
 public class MemoryManager implements MemoryController {
     
-    private HashMap<Integer, MemoryController> segmentControllers; // maps starting address to controller
+    private HashMap<Long, MemoryController> segmentControllers; // maps starting address to controller
     
-    private TreeSet<Integer> startAddresses; // good old tree for finding the right memory segment fast
+    private TreeSet<Long> startAddresses; // good old tree for finding the right memory segment fast
     
-    private HashMap<Integer, Integer> endAddresses; // maps starting address to ending address
+    private HashMap<Long, Long> endAddresses; // maps starting address to ending address
     
     /**
      * Create an empty MemoryManager
@@ -33,11 +33,11 @@ public class MemoryManager implements MemoryController {
      * @param end End address
      * @throws IllegalArgumentException if the specified range overlaps with an existing segment
      */
-    public void registerSegment(MemoryController mc, int start, int size) {
+    public void registerSegment(MemoryController mc, long start, long size) {
         // check for overlap
         if(this.segmentControllers.size() != 0) {
-            int closestSegmentStart = this.startAddresses.floor(start + size - 1),
-                closestSegmentEnd = this.endAddresses.get(closestSegmentStart);
+            long closestSegmentStart = this.startAddresses.floor(start + size - 1),
+                 closestSegmentEnd = this.endAddresses.get(closestSegmentStart);
             
             if(closestSegmentStart > start || closestSegmentEnd > start) {
                 throw new IllegalArgumentException("Memory segments cannot overlap");
@@ -55,7 +55,7 @@ public class MemoryManager implements MemoryController {
      * 
      * @param address Start address of the segment to remove
      */
-    public void removeSegment(int address) {
+    public void removeSegment(long address) {
         this.startAddresses.remove(address);
         this.endAddresses.remove(address);
         this.segmentControllers.remove(address);
@@ -69,8 +69,8 @@ public class MemoryManager implements MemoryController {
      * @throws IndexOutOfBoundsException when trying to access out-of-bounds memory
      * @throws NullPointerException if no segments have been registered
      */
-    private int getSegment(int address) {
-        Integer start = this.startAddresses.floor(address);
+    private long getSegment(long address) {
+        Long start = this.startAddresses.floor(address);
         
         if(start == null || this.endAddresses.get(start) < address) {
             throw new IndexOutOfBoundsException(String.format("Attempted to access non-registered address: %04X", address));
@@ -80,64 +80,64 @@ public class MemoryManager implements MemoryController {
     }
 
     @Override
-    public byte readByte(int address) {
-        int seg = getSegment(address);
+    public byte readByte(long address) {
+        long seg = getSegment(address);
         MemoryController sc = this.segmentControllers.get(seg);
         
         return sc.readByte(address - seg);
     }
     
     @Override
-    public short read2Bytes(int address) {
-        int seg = getSegment(address);
+    public short read2Bytes(long address) {
+        long seg = getSegment(address);
         MemoryController sc = this.segmentControllers.get(seg);
         
         return sc.read2Bytes(address - seg);
     }
     
     @Override
-    public int read3Bytes(int address) {
-        int seg = getSegment(address);
+    public int read3Bytes(long address) {
+        long seg = getSegment(address);
         MemoryController sc = this.segmentControllers.get(seg);
         
         return sc.read3Bytes(address - seg);
     }
     
     @Override
-    public int read4Bytes(int address) {
-        int seg = getSegment(address);
+    public int read4Bytes(long address) {
+        long seg = getSegment(address);
         MemoryController sc = this.segmentControllers.get(seg);
         
         return sc.read4Bytes(address - seg);
     }
 
     @Override
-    public void writeByte(int address, byte value) {
-        int seg = getSegment(address);
+    public void writeByte(long address, byte value) {
+        long seg = getSegment(address);
         MemoryController sc = this.segmentControllers.get(seg);
         
         sc.writeByte(address - seg, value);
     }
     
     @Override
-    public void write2Bytes(int address, short value) {
-        int seg = getSegment(address);
+    public void write2Bytes(long address, short value) {
+        long seg = getSegment(address);
         MemoryController sc = this.segmentControllers.get(seg);
         
         sc.write2Bytes(address - seg, value);
     }
     
     @Override
-    public void write3Bytes(int address, int value) {
-        int seg = getSegment(address);
+    public void write3Bytes(long address, int value) {
+        long seg = getSegment(address);
         MemoryController sc = this.segmentControllers.get(seg);
         
         sc.write3Bytes(address - seg, value);
     }
     
     @Override
-    public void write4Bytes(int address, int value) {
-        int seg = getSegment(address);
+    public void write4Bytes(long address, int value) {
+        long seg = getSegment(address);
         MemoryController sc = this.segmentControllers.get(seg);
         
         sc.write4Bytes(address - seg, value);
